fix(wallet-connect): crash when viewing connected dapps
