+++ .eslintrc.js
@@ -102,7 +102,6 @@ const allowedBarrelFiles = [
   'src/features/wallet-connect/stores/walletConnectRequestsStore.ts',
   'src/storage/index.ts',
   'src/framework/ui/styled-thing/index.tsx',
-  'src/features/wallet-connect/index.tsx',
 ];
 
 module.exports = {
