+++ src/hooks/useScanner.ts
@@ -17,7 +17,7 @@ import ethereumUtils from '@/utils/ethereumUtils';
 import haptics from '@/utils/haptics';
 import { logger, RainbowError } from '@/logger';
 import { checkPushNotificationPermissions } from '@/notifications/permissions';
-import { pair as pairWalletConnect } from '@/features/wallet-connect';
+import { pair as pairWalletConnect } from '@/features/wallet-connect/services/pair';
 import { getPoapAndOpenSheetWithQRHash, getPoapAndOpenSheetWithSecretWord } from '@/utils/poaps';
 
 export default function useScanner(enabled: boolean, onSuccess: () => unknown) {
